organize types: errors
