+++ types/errors/FieldError.d.ts
@@ -1,0 +1,11 @@
+export declare interface IFieldError {
+  /**
+   * Field of the request that the error describes
+   */
+  field: string;
+
+  /**
+   * Human readable description of the problem
+   */
+  message: string;
+}
